add limit to amount

+++ database/migrations/2024_04_21_185130_create_user_transactions_table.php
@@ -14,7 +14,7 @@ return new class () extends Migration {
             $table->id();
             $table->foreignId('user_id')->constrained();
             $table->foreignId('file_id')->nullable()->constrained();
-            $table->decimal('amount', 10, 2);
+            $table->double('amount');
             $table->string('operation');
             $table->string('type');
             $table->string('description')->nullable();
